+++ searx/botdetection/ip_lists.py
@@ -8,6 +8,10 @@ Method ``ip_lists``
 The ``ip_lists`` method implements IP :py:obj:`block- <block_ip>` and
 :py:obj:`pass-lists <pass_ip>`.
 
+
+Config
+~~~~~~
+
 .. code:: toml
 
    [botdetection.ip_lists]
@@ -21,6 +25,10 @@ The ``ip_lists`` method implements IP :py:obj:`block- <block_ip>` and
       '93.184.216.34', # IPv4 of example.org
       '257.1.1.1',     # invalid IP --> will be ignored, logged in ERROR class
    ]
+
+
+Implementations
+~~~~~~~~~~~~~~~
 
 """
 # pylint: disable=unused-argument
@@ -38,24 +46,11 @@ from ._helpers import logger
 
 logger = logger.getChild('ip_limit')
 
-SEARXNG_ORG = [
-    # https://github.com/searxng/searxng/pull/2484#issuecomment-1576639195
-    '167.235.158.251',  # IPv4 check.searx.space
-    '2a01:04f8:1c1c:8fc2::/64',  # IPv6 check.searx.space
-]
-"""Passlist of IPs from the SearXNG organization, e.g. `check.searx.space`."""
-
 
 def pass_ip(real_ip: IPv4Address | IPv6Address, cfg: config.Config) -> Tuple[bool, str]:
     """Checks if the IP on the subnet is in one of the members of the
     ``botdetection.ip_lists.pass_ip`` list.
     """
-
-    if cfg.get('botdetection.ip_lists.pass_searxng_org', default=True):
-        for net in SEARXNG_ORG:
-            net = ip_network(net, strict=False)
-            if real_ip.version == net.version and real_ip in net:
-                return True, f"IP matches {net.compressed} in SEARXNG_ORG list."
     return ip_is_subnet_of_member_in_list(real_ip, 'botdetection.ip_lists.pass_ip', cfg)
