+++ searx/botdetection/link_token.py
@@ -13,6 +13,9 @@ a ping by request a static URL.
 
    This method requires a redis DB and needs a HTTP X-Forwarded-For_ header.
 
+.. _X-Forwarded-For:
+   https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Forwarded-For
+
 To get in use of this method a flask URL route needs to be added:
 
 .. code:: python
@@ -31,10 +34,31 @@ And in the HTML template from flask a stylesheet link is needed (the value of
          href="{{ url_for('client_token', token=link_token) }}"
          type="text/css" />
 
-.. _X-Forwarded-For:
-   https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Forwarded-For
+
+Config
+~~~~~~
+
+.. code:: toml
+
+   [botdetection.link_token]
+   # Livetime (sec) of limiter's CSS token.
+   TOKEN_LIVE_TIME = 600
+
+   # Livetime (sec) of the ping-key from a client (request)
+   PING_LIVE_TIME = 3600
+
+   # Prefix of all ping-keys generated by link_token.get_ping_key
+   PING_KEY = 'botdetection.link_token.PING_KEY'
+
+   # Key for which the current token is stored in the DB
+   TOKEN_KEY = 'botdetection.link_token.TOKEN_KEY'
+
+
+Implementations
+~~~~~~~~~~~~~~~
 
 """
+
 from __future__ import annotations
 from ipaddress import (
     IPv4Network,
@@ -46,48 +70,47 @@ import string
 import random
 import flask
 
-from searx import logger
-from searx import redisdb
-from searx.redislib import secret_hash
+from . import ctx
+from .redislib import secret_hash
 
 from ._helpers import (
+    logger,
     get_network,
     get_real_ip,
 )
 
-TOKEN_LIVE_TIME = 600
-"""Livetime (sec) of limiter's CSS token."""
 
-PING_LIVE_TIME = 3600
-"""Livetime (sec) of the ping-key from a client (request)"""
+logger = logger.getChild('link_token')
 
-PING_KEY = 'SearXNG_limiter.ping'
+
+PING_KEY = 'botdetection.link_token.PING_KEY'
 """Prefix of all ping-keys generated by :py:obj:`get_ping_key`"""
 
-TOKEN_KEY = 'SearXNG_limiter.token'
+TOKEN_KEY = 'botdetection.link_token.TOKEN_KEY'
 """Key for which the current token is stored in the DB"""
 
-logger = logger.getChild('botdetection.link_token')
+
+def _cfg(name):
+    return ctx.cfg.get(f'botdetection.link_token.{name}')
 
 
 def is_suspicious(network: IPv4Network | IPv6Network, request: flask.Request, renew: bool = False):
     """Checks whether a valid ping is exists for this (client) network, if not
     this request is rated as *suspicious*.  If a valid ping exists and argument
     ``renew`` is ``True`` the expire time of this ping is reset to
-    :py:obj:`PING_LIVE_TIME`.
+    ``PING_LIVE_TIME``.
 
     """
-    redis_client = redisdb.client()
-    if not redis_client:
+    if not ctx.redis_client:
         return False
 
     ping_key = get_ping_key(network, request)
-    if not redis_client.get(ping_key):
+    if not ctx.redis_client.get(ping_key):
         logger.info("missing ping (IP: %s) / request: %s", network.compressed, ping_key)
         return True
 
     if renew:
-        redis_client.set(ping_key, 1, ex=PING_LIVE_TIME)
+        ctx.redis_client.set(ping_key, 1, ex=_cfg('PING_LIVE_TIME'))
 
     logger.debug("found ping for (client) network %s -> %s", network.compressed, ping_key)
     return False
@@ -96,22 +119,21 @@ def is_suspicious(network: IPv4Network | IPv6Network, request: flask.Request, re
 def ping(request: flask.Request, token: str):
     """This function is called by a request to URL ``/client<token>.css``.  If
     ``token`` is valid a :py:obj:`PING_KEY` for the client is stored in the DB.
-    The expire time of this ping-key is :py:obj:`PING_LIVE_TIME`.
+    The expire time of this ping-key is ``PING_LIVE_TIME``.
 
     """
-    from . import redis_client, cfg  # pylint: disable=import-outside-toplevel, cyclic-import
-
-    if not redis_client:
+    if not ctx.redis_client:
         return
     if not token_is_valid(token):
         return
 
     real_ip = ip_address(get_real_ip(request))
-    network = get_network(real_ip, cfg)
+    network = get_network(real_ip, ctx.cfg)
 
     ping_key = get_ping_key(network, request)
     logger.debug("store ping_key for (client) network %s (IP %s) -> %s", network.compressed, real_ip, ping_key)
-    redis_client.set(ping_key, 1, ex=PING_LIVE_TIME)
+
+    ctx.redis_client.set(ping_key, 1, ex=_cfg('PING_LIVE_TIME'))
 
 
 def get_ping_key(network: IPv4Network | IPv6Network, request: flask.Request) -> str:
@@ -137,19 +159,21 @@ def get_token() -> str:
     """Returns current token.  If there is no currently active token a new token
     is generated randomly and stored in the redis DB.
 
-    - :py:obj:`TOKEN_LIVE_TIME`
-    - :py:obj:`TOKEN_KEY`
+    Config:
+
+    - ``TOKEN_LIVE_TIME``
+    - ``TOKEN_KEY``
 
     """
-    redis_client = redisdb.client()
-    if not redis_client:
+    if not ctx.redis_client:
         # This function is also called when limiter is inactive / no redis DB
         # (see render function in webapp.py)
         return '12345678'
-    token = redis_client.get(TOKEN_KEY)
+    token_key = _cfg('TOKEN_KEY')
+    token = ctx.redis_client.get(token_key)
     if token:
         token = token.decode('UTF-8')
     else:
         token = ''.join(random.choice(string.ascii_lowercase + string.digits) for _ in range(16))
-        redis_client.set(TOKEN_KEY, token, ex=TOKEN_LIVE_TIME)
+        ctx.redis_client.set(token_key, token, ex=_cfg('TOKEN_LIVE_TIME'))
     return token
