+++ searx/limiter.toml
@@ -10,31 +10,25 @@ x_for = 1
 ipv4_prefix = 32
 ipv6_prefix = 48
 
-[botdetection.ip_limit]
+[botdetection.redis]
 
-# To get unlimited access in a local network, by default link-lokal addresses
-# (networks) are not monitored by the ip_limit
-filter_link_local = false
+# FQDN of a function definition. A function with which the DB keys of the Redis
+# DB are to be annonymized.
+secret_hash = 'searx.redislib.secret_hash'
 
-# activate link_token method in the ip_limit method
-link_token = false
+# A prefix to all keys stored by the botdetection in the redis DB
+REDIS_KEY_PREFIX = 'SearXNG_'
 
-[botdetection.ip_lists]
+[botdetection.link_token]
 
-# In the limiter, the ip_lists method has priority over all other methods -> if
-# an IP is in the pass_ip list, it has unrestricted access and it is also not
-# checked if e.g. the "user agent" suggests a bot (e.g. curl).
+# Livetime (sec) of limiter's CSS token.
+TOKEN_LIVE_TIME = 600
 
-block_ip = [
-  # '93.184.216.34',  # IPv4 of example.org
-  # '257.1.1.1',      # invalid IP --> will be ignored, logged in ERROR class
-]
+# Livetime (sec) of the ping-key from a client (request)
+PING_LIVE_TIME = 3600
 
-pass_ip = [
-  # '192.168.0.0/16',      # IPv4 private network
-  # 'fe80::/10'            # IPv6 linklocal / wins over botdetection.ip_limit.filter_link_local
-]
+# Prefix of all ping-keys generated by link_token.get_ping_key
+PING_KEY = 'SearXNG_limiter.ping'
 
-# Activate passlist of (hardcoded) IPs from the SearXNG organization,
-# e.g. `check.searx.space`.
-pass_searxng_org = true
+# Key for which the current token is stored in the DB
+TOKEN_KEY = 'SearXNG_limiter.token'
