+++ searx/search/checker/scheduler.py
@@ -18,7 +18,7 @@ import importlib
 from typing import Callable
 
 from searx.redisdb import client as get_redis_client
-from searx.redislib import lua_script_storage
+from searx.botdetection.redislib import lua_script_storage
 
 
 logger = logging.getLogger('searx.search.checker')
